+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,12 @@
+import StyledFooter from "./FooterStyles";
+
+const Footer = () => {
+  return (
+    <StyledFooter>
+      © 2022 | Przemysław Szalast for Capgemini React Academy | All Rights
+      Reserved.
+    </StyledFooter>
+  );
+};
+
+export default Footer;
