+++ src/components/MainHeader/MainHeaderStyles.ts
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+const StyledMainHeader = styled.header`
+  display: flex;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 5rem;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0 1rem;
+  background-color: #fff;
+`;
+
+export default StyledMainHeader;
