+++ src/components/SweetsList/SweetsList.tsx
@@ -1,0 +1,12 @@
+import { StyledMain, StyledSection } from "./SweetsListStyles";
+
+const SweetsList: React.FC = () => {
+  return (
+    <StyledMain>
+      <p>Order our delicious sweets now!</p>
+      <StyledSection>SweetsList</StyledSection>
+    </StyledMain>
+  );
+};
+
+export default SweetsList;
