+++ src/components/SweetsList/SweetsListStyles.ts
@@ -1,10 +1,9 @@
 import styled from "styled-components";
 import backgroundImage from "../assets/background.png";
 
-const Main = styled.main`
+export const StyledMain = styled.main`
   position: fixed;
   top: 5rem;
-
   background: linear-gradient(
       rgba(207, 187, 203, 0.81),
       rgba(207, 187, 203, 0.81)
@@ -23,23 +22,10 @@ const Main = styled.main`
   }
 `;
 
-const Section = styled.section`
-  max-width: 60rem;
-  /* width: 90%; */
+export const StyledSection = styled.section`
   margin: 3rem auto;
+  max-width: 60rem;
   text-align: center;
-
   border: 1px solid #fff;
   border-radius: 20px;
 `;
-
-const SweetsList: React.FC = () => {
-  return (
-    <Main>
-      <p>Order our delicious sweets now!</p>
-      <Section>SweetsList</Section>
-    </Main>
-  );
-};
-
-export default SweetsList;
