+++ src/index.css
@@ -9,3 +9,4 @@ html {
 
 body {
   height: 100%;
+}
